rag-confluence/rag-a-confluence/requirements.txt
requests>=2.32
python-dotenv>=1.0
markdownify>=0.13
beautifulsoup4>=4.12
google-genai>=0.3
streamlit>=1.38
